Decouple get_host_port tests from Flask
